--- src/main/java/tp/Command.java
@@ -1,13 +1,0 @@
-package tp;
-
-public abstract class Command {
-
-    public boolean isExit() {
-        return false;
-    }
-
-    public abstract void execute(DoctorList doctorList, PatientList patientList,
-                                   AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
-                                    PatientStorage patientStorage,
-                                    AppointmentStorage appointmentStorage) throws IHospitalException;
-}
