+++ src/main/java/tp/command/AddAppointmentCommand.java
@@ -1,5 +1,8 @@
 package tp;
 
+package tp.command;
+
+import tp.*;
 import tp.person.Doctor;
 import tp.person.Patient;
