+++ src/main/java/tp/command/Command.java
@@ -1,0 +1,15 @@
+package tp.command;
+
+import tp.*;
+
+public abstract class Command {
+
+    public boolean isExit() {
+        return false;
+    }
+
+    public abstract void execute(DoctorList doctorList, PatientList patientList,
+                                 AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
+                                 PatientStorage patientStorage,
+                                 AppointmentStorage appointmentStorage) throws IHospitalException;
+}
