+++ src/main/java/tp/command/DeletePatientCommand.java
@@ -1,12 +1,19 @@
-package tp;
+package tp.command;
 
-public class ListDoctorListCommand extends Command {
+import tp.*;
+
+public class DeletePatientCommand extends Command {
+    private final int index;
+
+    public DeletePatientCommand(int index) {
+        this.index = index;
+    }
 
     @Override
     public void execute(DoctorList doctorList, PatientList patientList,
                         AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
                         PatientStorage patientStorage,
                         AppointmentStorage appointmentStorage) throws IHospitalException {
-        System.out.print(doctorList);
+        patientList.deletePatient(index);
     }
 }
