+++ src/main/java/tp/command/ExitCommand.java
@@ -1,16 +1,18 @@
-package tp;
+package tp.command;
 
-public class SortAppointmentByTimeCommand extends Command {
+import tp.*;
 
-    public SortAppointmentByTimeCommand() {
+public class ExitCommand extends Command {
+    public ExitCommand() {
 
     }
+
 
     @Override
     public void execute(DoctorList doctorList, PatientList patientList,
                         AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
                         PatientStorage patientStorage,
                         AppointmentStorage appointmentStorage) throws IHospitalException {
-        appointmentList.sortByTime();
+        return;
     }
 }
