+++ src/main/java/tp/command/GetAppointmentsOfDoctorCommand.java
@@ -1,16 +1,19 @@
-package tp;
+package tp.command;
 
-public class ExitCommand extends Command {
-    public ExitCommand() {
+import tp.*;
 
+public class GetAppointmentsOfDoctorCommand extends Command {
+    private final String id;
+
+    public GetAppointmentsOfDoctorCommand(String id) {
+        this.id = id;
     }
-
 
     @Override
     public void execute(DoctorList doctorList, PatientList patientList,
                         AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
                         PatientStorage patientStorage,
                         AppointmentStorage appointmentStorage) throws IHospitalException {
-        return;
+        System.out.println(appointmentList.getAppointmentListOfDoctorById(id));
     }
 }
