+++ src/main/java/tp/command/ListAppointmentListCommand.java
@@ -1,12 +1,14 @@
-package tp;
+package tp.command;
 
-public class ListPatientListCommand extends Command {
+import tp.*;
+
+public class ListAppointmentListCommand extends Command {
 
     @Override
     public void execute(DoctorList doctorList, PatientList patientList,
                         AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
                         PatientStorage patientStorage,
                         AppointmentStorage appointmentStorage) throws IHospitalException {
-        System.out.print(patientList);
+        System.out.print(appointmentList);
     }
 }
