+++ src/main/java/tp/command/ListPatientListCommand.java
@@ -1,12 +1,14 @@
-package tp;
+package tp.command;
 
-public class ListAppointmentListCommand extends Command {
+import tp.*;
+
+public class ListPatientListCommand extends Command {
 
     @Override
     public void execute(DoctorList doctorList, PatientList patientList,
                         AppointmentList appointmentList, Ui ui, DoctorStorage doctorStorage,
                         PatientStorage patientStorage,
                         AppointmentStorage appointmentStorage) throws IHospitalException {
-        System.out.print(appointmentList);
+        System.out.print(patientList);
     }
 }
